fix: Fix the correction of transform
The transforms [wl_output::transform] named `flipped*` are composites of
flip across y-axis followed by rotations. grim attempts to invert the
transform applied to a screencopy frame, but the current implementation
treats `flipped*` as if composites of rotations followed by the flip.

This results wrong result for `flipped_90` and `flipped_180`. Note that
`flipped_180` is not affected as the two composites coincide.

This patch remedies it.

For more details, see
https://github.com/kenoss/sabiniwm/blob/a4364ca211f97b672f92fb4de78f570f74404b90/note/issue-transform.md
and #grim-transform-invert-wrong.

[wl_output::transform] https://wayland.freedesktop.org/docs/html/apa.html#protocol-spec-wl_output-enum-transform